Refactor compat module
