+++ certbot/client.py
@@ -448,7 +448,7 @@ class Client(object):
         """
         for path in cert_path, chain_path, fullchain_path:
             util.make_or_verify_dir(
-                os.path.dirname(path), 0o755, compat.os_geteuid(),
+                os.path.dirname(path), 0o755, compat.os.geteuid(),
                 self.config.strict_permissions)
