+++ certbot/main.py
@@ -1260,16 +1260,16 @@ def make_or_verify_needed_dirs(config):
 
     """
     util.set_up_core_dir(config.config_dir, constants.CONFIG_DIRS_MODE,
-                         compat.os_geteuid(), config.strict_permissions)
+                         compat.os.geteuid(), config.strict_permissions)
     util.set_up_core_dir(config.work_dir, constants.CONFIG_DIRS_MODE,
-                         compat.os_geteuid(), config.strict_permissions)
+                         compat.os.geteuid(), config.strict_permissions)
 
     hook_dirs = (config.renewal_pre_hooks_dir,
                  config.renewal_deploy_hooks_dir,
                  config.renewal_post_hooks_dir,)
     for hook_dir in hook_dirs:
         util.make_or_verify_dir(hook_dir,
-                                uid=compat.os_geteuid(),
+                                uid=compat.os.geteuid(),
                                 strict=config.strict_permissions)
 
 
@@ -1320,7 +1320,7 @@ def main(cli_args=sys.argv[1:]):
 
     # On windows, shell without administrative right cannot create symlinks required by certbot.
     # So we check the rights before continuing.
-    compat.raise_for_non_administrative_windows_rights(config.verb)
+    compat.misc.raise_for_non_administrative_windows_rights(config.verb)
 
     try:
         log.post_arg_parse_setup(config)
